split m_menu definitions to m_menudefs
idk if this is a good idea lel
